requirements.txt
djangorestframework==3.12.2
Django==3.1.7
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.7.0